[package] (Un)package GraphModules using corresponding Tracer
**Summary**
This commit updates how a `GraphModule` is packaged to pickle and
package the tracer used to produce said `GraphModule` (technically the
`Graph` it owns) and to unpackage it and use it to reconstruct the
`GraphModule` (as opposed to the vanilla tracer).

**Test Plan**
This commit adds `test_package_fx_custom_tracer`, a unit test that
traces a module with a custom tracer and packages it.

**Fixes**
This commit fixes #54573.

ghstack-source-id: e6ce2f2e4f8d849d8e7fe941b6c1fba410b7b9ab
Pull Request resolved: https://github.com/pytorch/pytorch/pull/57221